Do not treat bunders as node.js
When using this library under [Parcel](https://parceljs.org/), `isNode` is true even inside the browser. This is probably true for Webpack and other bundlers as well, since they all emulate the `module` system.

This version checks for `process.versions.node`, which is defined on ever version of Node.js since 0.2.0.